+++ tests/for_each_concurrent_limit.rs
@@ -1,24 +1,23 @@
-#![allow(deprecated)] // the deprecated iter_concurrent_limit! macro is still tested
-
 mod common;
 use core::time;
 use std::sync::atomic::AtomicUsize;
 
 use rayon::iter::{IntoParallelIterator, ParallelIterator};
 
-use rayon_iter_concurrent_limit::iter_concurrent_limit;
+use rayon_iter_concurrent_limit::ParallelIteratorConcurrentLimit;
 
 use common::{calc_active_operations, incr_active_operations};
 
 const DUR: time::Duration = core::time::Duration::from_millis(10);
 
-fn iter_concurrent_limit_filter_map(concurrent_limit: usize) {
+fn for_each_concurrent_limit(concurrent_limit: usize) {
     let threads_active = AtomicUsize::new(0);
     let threads_active_max = AtomicUsize::new(0);
     let threads_active_inner = AtomicUsize::new(0);
     let threads_active_inner_max = AtomicUsize::new(0);
-    let output =
-        iter_concurrent_limit!(concurrent_limit, 0..10, filter_map, |i| -> Option<usize> {
+    (0..10)
+        .into_par_iter()
+        .for_each_concurrent_limit(concurrent_limit, |_| {
             incr_active_operations(&threads_active);
             std::thread::sleep(DUR);
             (0..10).into_par_iter().for_each(|_| {
@@ -27,20 +26,7 @@ fn iter_concurrent_limit_filter_map(concurrent_limit: usize) {
                 calc_active_operations(&threads_active_inner, &threads_active_inner_max);
             });
             calc_active_operations(&threads_active, &threads_active_max);
-            if i % 2 == 0 {
-                Some(i)
-            } else {
-                None
-            }
-        })
-        .collect::<Vec<_>>();
-    assert_eq!(
-        output,
-        (0..10)
-            .into_iter()
-            .filter_map(|i| if i % 2 == 0 { Some(i) } else { None })
-            .collect::<Vec<_>>()
-    );
+        });
     assert_eq!(threads_active_max.into_inner(), concurrent_limit);
     if cfg!(not(feature = "ci")) {
         assert!(threads_active_inner_max.into_inner() > concurrent_limit);
@@ -48,17 +34,17 @@ fn iter_concurrent_limit_filter_map(concurrent_limit: usize) {
 }
 
 #[test]
-fn iter_concurrent_limit_filter_map_1() {
-    iter_concurrent_limit_filter_map(1);
+fn for_each_concurrent_limit_1() {
+    for_each_concurrent_limit(1);
 }
 
 #[test]
-fn iter_concurrent_limit_filter_map_2() {
-    iter_concurrent_limit_filter_map(2);
+fn for_each_concurrent_limit_2() {
+    for_each_concurrent_limit(2);
 }
 
 #[cfg_attr(feature = "ci", ignore)]
 #[test]
-fn iter_concurrent_limit_filter_map_4() {
-    iter_concurrent_limit_filter_map(4);
+fn for_each_concurrent_limit_4() {
+    for_each_concurrent_limit(4);
 }